Add no_std support for embedded targets
- Switch lib.rs to #![cfg_attr(not(feature = "std"), no_std)] with
  extern crate alloc for heap types (Vec, String, format!)
- Make regex and rustyline optional deps under "std" feature;
  mqjs binary requires-features = ["std"]
- Add libm = "0.2" for math functions in no_std context
- Replace std::mem/ptr/core usages with core:: equivalents in
  allocator.rs and ops.rs
- Downgrade edition 2024 → 2021 for broader toolchain compatibility
- Export NativeFn from crate root for embedder use

+++ src/runtime/function.rs
@@ -6,6 +6,7 @@
 //! - Function bytecode
 
 use crate::value::Value;
+use alloc::{string::String, vec::Vec, vec, format, string::ToString};
 
 /// Maximum number of function arguments
 pub const MAX_ARGS: u16 = 65535;
@@ -118,6 +119,7 @@ pub struct CaptureInfo {
     pub is_local: bool,
 }
 
+#[repr(align(8))]
 pub struct FunctionBytecode {
     /// Function name (for debugging)
     pub name: Option<String>,
@@ -404,7 +406,7 @@ impl FunctionBytecode {
         let mut constants = Vec::with_capacity(const_count);
         for _ in 0..const_count {
             let raw = read_u64(&mut pos)? as usize;
-            constants.push(Value(crate::value::RawValue(raw)));
+            constants.push(Value(crate::value::RawValue(raw as u64)));
         }
 
         // String constants
